pipeline speedups, zarr fixes, tutorial_demo visualization cells
- CNMFeParams: add sample_frames, skip_first_deconv, bg_tsub params
- temporal.py: deconvolve=False flag for first BCD pass (NNLS only)
- background.py: tsub temporal subsampling for ring-background W solve
- io.py: uint8 default dtype, blosc lz4+bitshuffle compression
- concat_avis_to_zarr.py / convert_to_zarr.py: uint8 defaults
- tutorial_demo.ipynb: real-video workflow with lazy zarr load,
  MC AVI export, 3-panel comparison movie (MC | overlay | recon),
  fix zarr[:]  / np.asarray(z[idx]) indexing (avoids per-frame decompression)

+++ convert_to_zarr.py
@@ -32,7 +32,7 @@ def main() -> None:
             print(f"  {avi_path.name}: zarr already exists, skipping")
             continue
         print(f"  {avi_path.name} → {zarr_path.name} …", end=" ", flush=True)
-        z = avi_to_zarr(avi_path, zarr_path, chunk_t=100, grayscale=True, dtype="float32")
+        z = avi_to_zarr(avi_path, zarr_path, chunk_t=100, grayscale=True)
         print(f"shape={z.shape}  chunks={z.chunks}")
 
     print("\nDone. Open zarrs lazily with  cnmfe.io.open_zarr('demo_movies/<name>.zarr')")
